+++ app/src/main/java/com/iboalali/basicrootchecker/ui/about/AboutScreen.kt
@@ -36,6 +36,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLayoutDirection
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.PreviewDynamicColors
@@ -97,6 +98,7 @@ fun AboutScreen(onNavigateBack: () -> Unit) {
         val bottomPadding = innerPadding.calculateBottomPadding()
         Column(
             modifier = Modifier
+                .testTag("about_list")
                 .fillMaxSize()
                 .padding(contentPadding)
                 .verticalScroll(rememberScrollState())